+++ Sources/AlgoliaSearch-Client/FilterBuilder/Auxiliary/Attribute.swift
