+++ Sources/AlgoliaSearch-Client/FilterBuilder/Auxiliary/FilterOperators.swift
