+++ Sources/AlgoliaSearch-Client/FilterBuilder/Auxiliary/NumericOperator.swift
