+++ Sources/AlgoliaSearch-Client/FilterBuilder/Filters/Filter.swift
