+++ Sources/AlgoliaSearch-Client/FilterBuilder/Filters/FilterFacet.swift
