+++ Sources/AlgoliaSearch-Client/FilterBuilder/Filters/FilterNumeric.swift
