+++ Sources/AlgoliaSearch-Client/FilterBuilder/Proxies/AndGroupProxy.swift
