+++ Sources/AlgoliaSearch-Client/FilterBuilder/Proxies/GroupProxy.swift
