+++ Sources/AlgoliaSearch-Client/FilterBuilder/Proxies/OrGroupProxy.swift
