+++ Sources/AlgoliaSearch-Client/FilterBuilder/Proxies/SpecializedAndGroupProxy.swift
